finished day 19 with some help
Basically, had a hint to say it was factoring.
This can be done in o(n) time. It was helpful to find
the inner loop and then just print how the registers were changing.
This showed that the first register was always
incremeneted by the third as soon as reg 1 * reg3 = reg 5 (or something)

This could be verified by setting reg5 to 16 which is easy to solve